osx: Use ssize_t for readlinkat()
It looks like OSX 10.10 finally has these functions.

+++ src/compat/macosx.h
@@ -24,6 +24,7 @@
 #include <stdlib.h>
 #include <sys/stat.h>
 #include <dirent.h>
+#include <unistd.h>
 
 /* MacOSX 10.6 does not have *at() functions */
 DIR *fdopendir(int fd);
@@ -33,7 +34,7 @@ int fchownat(int dirfd, const char *pathname, uid_t owner, gid_t group, int flag
 int fstatat(int dirfd, const char *pathname, struct stat *buf, int flags);
 int mkdirat(int dirfd, const char *pathname, mode_t mode);
 int openat(int dirfd, const char *pathname, int flags, ...);
-int readlinkat(int dirfd, const char *pathname, char *buf, size_t bufsiz);
+ssize_t readlinkat(int dirfd, const char *pathname, char *buf, size_t bufsiz);
 int renameat(int olddirfd, const char *oldpath, int newdirfd, const char *newpath);
 int symlinkat(const char *oldpath, int newdirfd, const char *newpath);
 int unlinkat(int dirfd, const char *pathname, int flags);
